requirements.txt
PyGithub==2.1.1
google-generativeai==0.3.2
